scripts/python/parser/requirements.txt
Package         Version  
--------------- ---------
certifi         2020.12.5
chardet         4.0.0    
click           7.1.2    
fastapi         0.63.0   
h11             0.12.0   
httptools       0.1.1    
idna            2.10     
pdftotext       2.1.5    
pip             20.0.2   
pkg-resources   0.0.0    
psycopg2-binary 2.8.6    
pydantic        1.7.3    
python-dotenv   0.15.0   
PyYAML          5.4.1    
requests        2.25.1   
setuptools      44.0.0   
SQLAlchemy      1.3.23   
starlette       0.13.6   
urllib3         1.26.3   
uvicorn         0.13.3   
uvloop          0.14.0   
watchgod        0.6      
websockets      8.1      
